+++ app/src/main/java/com/billdoerr/android/carputer/taskutils/TaskResult.java
@@ -1,0 +1,7 @@
+package com.billdoerr.android.carputer.taskutils;
+
+public class TaskResult {
+    public TaskRequest request;
+    public String cmdResult;
+    public String exception;
+}