+++ app/src/main/java/com/billdoerr/android/carputer/utils/NodeUtils.java
@@ -4,6 +4,7 @@ import android.util.Log;
 
 import com.jcraft.jsch.ChannelExec;
 import com.jcraft.jsch.JSch;
+import com.jcraft.jsch.JSchException;
 import com.jcraft.jsch.Session;
 
 import java.io.BufferedReader;
@@ -19,6 +20,8 @@ import java.util.Properties;
 public class NodeUtils {
 
     private static final String TAG = "NodeUtils";
+
+    private static final int TIMEOUT = 10*1000;     //  Time in milliseconds
 
     //  Hosts info
     private String mIP;
@@ -46,35 +49,55 @@ public class NodeUtils {
      * @return String: Result of command being executed.
      * @throws Exception Exception: Exception being thrown.
      */
-    public String executeRemoteCommand(String cmd) throws Exception {
+    public String executeRemoteCommand(String cmd) throws JSchException {
         JSch jsch = new JSch();
-        Session session = jsch.getSession(mUser, mIP, Integer.parseInt(mPort));
-        session.setPassword(mPwd);
+        Session session = null;
+        ChannelExec channelssh = null;
+        ByteArrayOutputStream output = null;
 
-        // Avoid asking for key confirmation
-        Properties prop = new Properties();
-        prop.put("StrictHostKeyChecking", "no");
+        try {
+            session = jsch.getSession(mUser, mIP, Integer.parseInt(mPort));
+            session.setPassword(mPwd);
 
-        session.setConfig(prop);
-        session.connect();
+            // Avoid asking for key confirmation
+            Properties prop = new Properties();
+            prop.put("StrictHostKeyChecking", "no");
 
-        // SSH Channel
-        ChannelExec channelssh = (ChannelExec) session.openChannel("exec");
-        ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        channelssh.setOutputStream(baos);
+            session.setConfig(prop);
+            session.connect();
+//            session.connect(TIMEOUT);
 
-        // Execute command
-        channelssh.setCommand(cmd);
-        channelssh.connect();
+            // SSH Channel
+            channelssh = (ChannelExec) session.openChannel("exec");
+            output = new ByteArrayOutputStream();
+            channelssh.setOutputStream(output);
 
-        //  Need delay
-        Thread.sleep(1000);
+            // Execute command
+            channelssh.setCommand(cmd);
+            channelssh.connect();
 
-        channelssh.disconnect();
+            //  Need delay
+            Thread.sleep(1000);
 
-        Log.d(TAG, "Execute Cmd Results...\n=====================\n" + baos.toString());
+        } catch (JSchException e) {
+            Log.e(TAG, e.toString());
+            throw e;
+        } catch (Exception e){
+            Log.e(TAG, e.toString());
+        } finally{
+            if (channelssh != null) {
+                channelssh.disconnect();
+            }
+            session.disconnect();
+        }
 
-        return baos.toString();
+        if (output != null) {
+            Log.d(TAG, "Execute Cmd Results...\n=====================\n" + output.toString());
+            return output.toString();
+        } else {
+            Log.d(TAG, "Execute Cmd Results...\n=====================\n" + "no results!");
+            return "";
+        }
     }
 
 
@@ -103,7 +126,7 @@ public class NodeUtils {
 
             while ((line = br.readLine()) != null) {
                 reply = reply + System.getProperty("line.separator") + line;
-                Log.d(TAG, reply);
+                Log.i(TAG, reply);
             }
         }
         catch (IOException e)